Refactor: 어드민 Qna 리팩토링 -> DTO 타입 일부 변경, 코드 스타일링, QnaService @Transactional 추가

+++ src/main/java/com/tbp/honeyjar/admin/dto/qna/QnaListResponseDto.java
@@ -5,8 +5,8 @@ import lombok.Data;
 @Data
 public class QnaListResponseDto {
 
-    private long id;
+    private Long id;
     private String title;
     private String post;
-    private int categoryId;
+    private Long categoryId;
 }
