+++ src/main/java/com/tbp/honeyjar/admin/dto/qna/QnaSaveRequestDto.java
@@ -7,11 +7,11 @@ import java.time.LocalDateTime;
 @Data
 public class QnaSaveRequestDto {
 
-    private long id;
+    private Long id;
     private String title;
     private String post;
-    private int categoryId;
-    private int userId;
+    private Long categoryId;
+    private Long userId;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 }
